refactor: 참여자 ID participantValidator로 이동, 관련 코드 수정
