+++ src/middlewares/validation/badgeValidator.js
@@ -1,0 +1,4 @@
+import { groupIdParamSchema, validateGroupIdParam } from '../validation/groupValidator.js';
+
+export { validateGroupIdParam };
+// 필요하면 badge 고유의 추가 스키마 및 미들웨어 작성 가능
